pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "job-applier"
version = "0.1.0"
description = "Autonomous background agent for job applications"
requires-python = ">=3.11"
dependencies = [
    # Orchestration
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.3.0",

    # Browser automation
    "browser-use>=0.1.0",
    "playwright>=1.44.0",

    # Web scraping
    "crawl4ai>=0.4.0",
    "httpx>=0.28.0",

    # LLM
    "openai>=1.50.0",
    "tiktoken>=0.7.0",

    # Backend API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "python-multipart>=0.0.12",
    "websockets>=13.0",

    # Database
    "sqlalchemy[asyncio]>=2.0.0",
    "aiosqlite>=0.20.0",
    "alembic>=1.13.0",

    # Scheduling
    "apscheduler>=3.10.0",

    # Config and utils
    "pydantic>=2.8.0",
    "pydantic-settings>=2.4.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "jinja2>=3.1.0",
    "rich>=13.0.0",
    "click>=8.1.0",

    # PDF generation (cover letters)
    "weasyprint>=62.0",
    "markdown2>=2.4.0",

    # Testing
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.28.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
python_files = ["test_*.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
